General update

+++ src/tools/extractors/binExtractor.ts
@@ -205,8 +205,13 @@ export default function BinExtractor(binFilePath: string, outputDirPath: string,
     if (exportDecompressedBin) {
         logger.warn("Exporting the decompressed bin file, this may take a a bit more time..");
 
-        const outputFilePath = path.join(outputDirPath, "decompressed_" + path.basename(binFilePath));
-        fs.writeFileSync(outputFilePath, Buffer.concat(dataBlocks));
+        const outputFilePath = path.join(outputDirPath, path.basename(binFilePath, ".bin") + ".decomp.bin");
+
+        if (!fs.existsSync(outputFilePath)) {
+            fs.writeFileSync(outputFilePath, Buffer.concat(dataBlocks));
+        } else {
+            logger.warn(`The decompressed version of ${path.basename(binFilePath)} already exists, skipping..`);
+        }
     }
 
     // Closing the file from the cache
